fix ISO C errors

+++ src/dslib/datastr.c
@@ -37,27 +37,27 @@ char *str_data_to_string(void *arg) {
 char *location_str_data_to_string(void *arg) {
   Location *location = (Location *)arg;
   char *data = (char *)location->data;
-  int cost = location->cost == UINT_MAX ? -1 : location->cost;
+  long cost = location->cost == UINT_MAX ? -1 : (long)location->cost;
   char *buffer = malloc(50);
-  snprintf(buffer, 50, "[%d]-%d-%s", location->id, cost, data);
+  snprintf(buffer, 50, "[%d]-%ld-%s", location->id, cost, data);
   return buffer;
 }
 
 char *location_char_data_to_string(void *arg) {
   Location *location = (Location *)arg;
   char data = *(char *)arg;
-  int cost = location->cost == UINT_MAX ? -1 : location->cost;
+  long cost = location->cost == UINT_MAX ? -1 : (long)location->cost;
   char *buffer = malloc(50);
-  snprintf(buffer, 50, "[%d]-%d-%c", location->id, cost, data);
+  snprintf(buffer, 50, "[%d]-%ld-%c", location->id, cost, data);
   return buffer;
 }
 
 char *location_int_data_to_string(void *arg) {
   Location *location = (Location *)arg;
   int data = *(int *)location->data;
-  int cost = location->cost == UINT_MAX ? -1 : location->cost;
+  long cost = location->cost == UINT_MAX ? -1 : (long)location->cost;
   char *buffer = malloc(50);
-  snprintf(buffer, 50, "[%d]-%d-%d", location->id, cost, data);
+  snprintf(buffer, 50, "[%d]-%ld-%d", location->id, cost, data);
   return buffer;
 }
